rename main_plot module to plotter, move bars module to plotter module
