--- pytripgui/canvas_vc/main_plot/images/__init__.py
@@ -1,5 +1,0 @@
-from pytripgui.canvas_vc.main_plot.images.ctx_image import CtxImage
-from pytripgui.canvas_vc.main_plot.images.dose_image import DoseImage
-from pytripgui.canvas_vc.main_plot.images.let_image import LetImage
-
-__all__ = ['CtxImage', 'DoseImage', 'LetImage']
