+++ pytripgui/canvas_vc/plotter/__init__.py
@@ -1,5 +1,5 @@
-from pytripgui.canvas_vc.main_plot.coordinate_info import CoordinateInfo
-from pytripgui.canvas_vc.main_plot.plotting_manager import PlottingManager
+from pytripgui.canvas_vc.plotter.coordinate_info import CoordinateInfo
+from pytripgui.canvas_vc.plotter.plotting_manager import PlottingManager
 
 # from https://docs.python.org/3/tutorial/modules.html
 # if a package's __init__.py code defines a list named __all__,
