+++ pytripgui/canvas_vc/plotter/bars/__init__.py
@@ -1,7 +1,4 @@
-from pytripgui.canvas_vc.bars.ctx_bar import CtxBar
-from pytripgui.canvas_vc.bars.dos_bar import DosBar
-from pytripgui.canvas_vc.bars.let_bar import LetBar
-from pytripgui.canvas_vc.bars.projection_enum import BarProjection
+from pytripgui.canvas_vc.plotter.bars.projection_enum import BarProjection
 
 # from https://docs.python.org/3/tutorial/modules.html
 # if a package's __init__.py code defines a list named __all__,
