+++ pytripgui/canvas_vc/plotter/bars/let_bar.py
@@ -1,7 +1,7 @@
 from matplotlib.projections import register_projection
 
-from pytripgui.canvas_vc.bars.bar_base import BarBase
-from pytripgui.canvas_vc.bars.projection_enum import BarProjection
+from pytripgui.canvas_vc.plotter.bars.bar_base import BarBase
+from pytripgui.canvas_vc.plotter.bars.projection_enum import BarProjection
 
 
 class LetBar(BarBase):
