+++ pytripgui/canvas_vc/plotter/images/patient_image_base.py
